moved test into start-manifest e2e project
