--- e2e/react-start/css-modules/src/styles/shared-widget.module.css
@@ -1,19 +1,0 @@
-.widget {
-  background: rgb(255, 247, 237);
-  border: 2px solid rgb(249, 115, 22);
-  border-radius: 14px;
-  padding: 20px;
-}
-
-.title {
-  color: rgb(154, 52, 18);
-  font-size: 20px;
-  font-weight: 700;
-  margin-bottom: 8px;
-}
-
-.content {
-  color: rgb(194, 65, 12);
-  font-size: 14px;
-  line-height: 1.5;
-}
