+++ e2e/react-start/start-manifest/src/components/AppShell.tsx
@@ -5,6 +5,8 @@ const ROUTES = linkOptions([
   { to: '/', label: 'home' },
   { to: '/a', label: '/a' },
   { to: '/b', label: '/b' },
+  { to: '/lazy-css-static', label: '/lazy-css-static' },
+  { to: '/lazy-css-lazy', label: '/lazy-css-lazy' },
   { to: '/r1', label: '/r1' },
   { to: '/r2', label: '/r2' },
   { to: '/shared-a', label: '/shared-a' },
