+++ e2e/react-start/start-manifest/src/components/SharedWidgetLazy.tsx
@@ -1,4 +1,4 @@
-import { SharedWidget } from './shared-widget'
+import { SharedWidget } from './SharedWidget'
 
 export default function SharedWidgetLazy() {
   return <SharedWidget />
