+++ e2e/react-start/start-manifest/src/routeTree.gen.ts
@@ -18,6 +18,8 @@ import { Route as R4RouteImport } from './routes/r4'
 import { Route as R3RouteImport } from './routes/r3'
 import { Route as R2RouteImport } from './routes/r2'
 import { Route as R1RouteImport } from './routes/r1'
+import { Route as LazyCssStaticRouteImport } from './routes/lazy-css-static'
+import { Route as LazyCssLazyRouteImport } from './routes/lazy-css-lazy'
 import { Route as BRouteImport } from './routes/b'
 import { Route as ARouteImport } from './routes/a'
 import { Route as IndexRouteImport } from './routes/index'
@@ -67,6 +69,16 @@ const R1Route = R1RouteImport.update({
   path: '/r1',
   getParentRoute: () => rootRouteImport,
 } as any)
+const LazyCssStaticRoute = LazyCssStaticRouteImport.update({
+  id: '/lazy-css-static',
+  path: '/lazy-css-static',
+  getParentRoute: () => rootRouteImport,
+} as any)
+const LazyCssLazyRoute = LazyCssLazyRouteImport.update({
+  id: '/lazy-css-lazy',
+  path: '/lazy-css-lazy',
+  getParentRoute: () => rootRouteImport,
+} as any)
 const BRoute = BRouteImport.update({
   id: '/b',
   path: '/b',
@@ -87,6 +99,8 @@ export interface FileRoutesByFullPath {
   '/': typeof IndexRoute
   '/a': typeof ARoute
   '/b': typeof BRoute
+  '/lazy-css-lazy': typeof LazyCssLazyRoute
+  '/lazy-css-static': typeof LazyCssStaticRoute
   '/r1': typeof R1Route
   '/r2': typeof R2Route
   '/r3': typeof R3Route
@@ -101,6 +115,8 @@ export interface FileRoutesByTo {
   '/': typeof IndexRoute
   '/a': typeof ARoute
   '/b': typeof BRoute
+  '/lazy-css-lazy': typeof LazyCssLazyRoute
+  '/lazy-css-static': typeof LazyCssStaticRoute
   '/r1': typeof R1Route
   '/r2': typeof R2Route
   '/r3': typeof R3Route
@@ -116,6 +132,8 @@ export interface FileRoutesById {
   '/': typeof IndexRoute
   '/a': typeof ARoute
   '/b': typeof BRoute
+  '/lazy-css-lazy': typeof LazyCssLazyRoute
+  '/lazy-css-static': typeof LazyCssStaticRoute
   '/r1': typeof R1Route
   '/r2': typeof R2Route
   '/r3': typeof R3Route
@@ -132,6 +150,8 @@ export interface FileRouteTypes {
     | '/'
     | '/a'
     | '/b'
+    | '/lazy-css-lazy'
+    | '/lazy-css-static'
     | '/r1'
     | '/r2'
     | '/r3'
@@ -146,6 +166,8 @@ export interface FileRouteTypes {
     | '/'
     | '/a'
     | '/b'
+    | '/lazy-css-lazy'
+    | '/lazy-css-static'
     | '/r1'
     | '/r2'
     | '/r3'
@@ -160,6 +182,8 @@ export interface FileRouteTypes {
     | '/'
     | '/a'
     | '/b'
+    | '/lazy-css-lazy'
+    | '/lazy-css-static'
     | '/r1'
     | '/r2'
     | '/r3'
@@ -175,6 +199,8 @@ export interface RootRouteChildren {
   IndexRoute: typeof IndexRoute
   ARoute: typeof ARoute
   BRoute: typeof BRoute
+  LazyCssLazyRoute: typeof LazyCssLazyRoute
+  LazyCssStaticRoute: typeof LazyCssStaticRoute
   R1Route: typeof R1Route
   R2Route: typeof R2Route
   R3Route: typeof R3Route
@@ -251,6 +277,20 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof R1RouteImport
       parentRoute: typeof rootRouteImport
     }
+    '/lazy-css-static': {
+      id: '/lazy-css-static'
+      path: '/lazy-css-static'
+      fullPath: '/lazy-css-static'
+      preLoaderRoute: typeof LazyCssStaticRouteImport
+      parentRoute: typeof rootRouteImport
+    }
+    '/lazy-css-lazy': {
+      id: '/lazy-css-lazy'
+      path: '/lazy-css-lazy'
+      fullPath: '/lazy-css-lazy'
+      preLoaderRoute: typeof LazyCssLazyRouteImport
+      parentRoute: typeof rootRouteImport
+    }
     '/b': {
       id: '/b'
       path: '/b'
@@ -279,6 +319,8 @@ const rootRouteChildren: RootRouteChildren = {
   IndexRoute: IndexRoute,
   ARoute: ARoute,
   BRoute: BRoute,
+  LazyCssLazyRoute: LazyCssLazyRoute,
+  LazyCssStaticRoute: LazyCssStaticRoute,
   R1Route: R1Route,
   R2Route: R2Route,
   R3Route: R3Route,
