+++ e2e/react-start/start-manifest/src/routes/lazy-css-lazy.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { createFileRoute } from '@tanstack/react-router'
 
 const LazySharedWidget = React.lazy(
-  () => import('~/components/shared-widget-lazy'),
+  () => import('~/components/SharedWidgetLazy'),
 )
 
 export const Route = createFileRoute('/lazy-css-lazy')({
@@ -11,7 +11,7 @@ export const Route = createFileRoute('/lazy-css-lazy')({
 
 function LazyCssLazyRoute() {
   return (
-    <div>
+    <section>
       <h1 data-testid="lazy-css-lazy-heading">Lazy CSS Repro - Lazy Route</h1>
       <p>
         This route renders the same widget through React.lazy so the CSS only
@@ -23,6 +23,6 @@ function LazyCssLazyRoute() {
       >
         <LazySharedWidget />
       </React.Suspense>
-    </div>
+    </section>
   )
 }
